clean: Move debug to separate package
